// src/lib/types/elevenlabs/voices.ts
// Generated by https://quicktype.io

export interface Voices {
  voices: Voice[];
}

export interface Voice {
  voice_id: string;
  name: string;
  samples: Sample[] | null;
  category: Category;
  fine_tuning: FineTuning;
  labels: Labels;
  description: null | string;
  preview_url: string | null;
  available_for_tiers: unknown[];
  settings: null;
  sharing: null;
  high_quality_base_model_ids: HighQualityBaseModelID[];
  safety_control: null;
  voice_verification: VoiceVerification;
}

export enum Category {
  Cloned = "cloned",
  Generated = "generated",
  Premade = "premade"
}

export interface FineTuning {
  is_allowed_to_fine_tune: boolean;
  finetuning_state: FinetuningState;
  verification_failures: unknown[];
  verification_attempts_count: number;
  manual_verification_requested: boolean;
  language: null | string;
  finetuning_progress: FinetuningProgress;
  message: null;
  dataset_duration_seconds: null;
  verification_attempts: null;
  slice_ids: null;
  manual_verification: null;
}

export interface FinetuningProgress {}

export enum FinetuningState {
  FineTuned = "fine_tuned",
  NotStarted = "not_started"
}

export enum HighQualityBaseModelID {
  ElevenMultilingualV1 = "eleven_multilingual_v1",
  ElevenTurboV2 = "eleven_turbo_v2"
}

export interface Labels {
  accent?: string;
  description?: string;
  age?: Age;
  gender?: Gender;
  "use case"?: string;
  "description "?: string;
  featured?: string;
  usecase?: string;
}

export enum Age {
  AgeMiddleAged = "middle_aged",
  MiddleAged = "middle aged",
  Old = "old",
  Young = "young"
}

export enum Gender {
  Female = "female",
  Male = "male"
}

export interface Sample {
  sample_id: string;
  file_name: string;
  mime_type: MIMEType;
  size_bytes: number;
  hash: string;
}

export enum MIMEType {
  AudioMPEG = "audio/mpeg"
}

export interface VoiceVerification {
  requires_verification: boolean;
  is_verified: boolean;
  verification_failures: unknown[];
  verification_attempts_count: number;
  language: null;
  verification_attempts: null;
}